modify stats components
